tools: Store return values from fread() in a size_t
In annexb_read, the res variable was used both for the return
from fread() (which is size_t, unsigned) and from leb128 (which
returns a signed int); remove the return value assignment
altogether as it wasn't used.

+++ tools/input/ivf.c
@@ -55,7 +55,7 @@ static int64_t rl64(const uint8_t *const p) {
 static int ivf_open(IvfInputContext *const c, const char *const file,
                     unsigned fps[2], unsigned *const num_frames)
 {
-    int res;
+    size_t res;
     uint8_t hdr[32];
 
     memset(c, 0, sizeof(*c));
@@ -97,7 +97,7 @@ static int ivf_open(IvfInputContext *const c, const char *const file,
 static int ivf_read(IvfInputContext *const c, Dav1dData *const buf) {
     uint8_t data[8];
     uint8_t *ptr;
-    int res;
+    size_t res;
 
     const int64_t off = ftello(c->f);
     if ((res = fread(data, 4, 1, c->f)) != 1)
